requirements.txt
langchain
landchain-experimental
pandas
openai
streamlit-chat
datetime
sqlalchemy
pyodbc
snowflake-connector-python
snowflake-sqlalchemy